algorithm/tree/102_Binary_Tree_Level_Order_Traversal/solution.go
package Binary_Tree_Level_Order_Traversal

import . "github.com/sevenger/gopher30days/algorithm"

func levelOrder(root *TreeNode) [][]int {
	if root == nil {
		return nil
	}

	var res [][]int
	queue := []*TreeNode{root}
	for len(queue) > 0 {
		l := len(queue)
		list := make([]int, 0, l)
		for i := 0; i < l; i++ {
			node := queue[0]
			queue = queue[1:]
			list = append(list, node.Val)
			if node.Left != nil {
				queue = append(queue, node.Left)
			}
			if node.Right != nil {
				queue = append(queue, node.Right)
			}
		}
		res = append(res, list)
	}
	return res
}
